LayoutTests/fast/css/font-face-cache-bug.html

<html>
  <script src="../js/resources/js-test-pre.js"></script>
  <script>
    description('Test for <a href="http://webkit.org/b/73520">WebKit bug 73520</a>');
    window.jsTestIsAsync = true;
    function runTest()
    {
        testDiv = document.getElementById('test');
        shouldBeEqualToString('window.getComputedStyle(testDiv, 0).width', '48px');
        finishJSTest();
    }
  </script>
  <head>
    <style type="text/css">
      body {
        font-family: serif;
      }
      @font-face{
        font-family: 'ahem';
        src: url('../../resources/Ahem.ttf');
      }
      #test {
        position: absolute
      }
      #test span {
        font-family: ahem;
      }
    </style>
  </head>
  <body onload="setTimeout('runTest()', 100)">
    <div id=test>
      <a href="#">
        <span>aaa</span>
      </a>
    </div>
    <script src="../js/resources/js-test-post.js"></script>
  </body>
</html>
